Fixing re-connecting issue in the harness
