+++ integration-tests/environments/browser/runner.ts
@@ -1,6 +1,4 @@
 import puppeteer from 'puppeteer';
-// import vite from "vite";
-// import { Server as MochaRemoteServer } from "mocha-remote-server";
 
 const DEV_TOOLS = process.env.DEV_TOOLS === "true" || process.env.DEV_TOOLS === "1";
 
@@ -16,10 +14,4 @@ page.on('console', msg => {
     }
 });
 
-await new Promise(resolve => setTimeout(resolve, 1000));
 await page.goto("http://localhost:5173/");
-
-// Keeping the process alive for the tests to complete
-setInterval(() => {
-    console.log("Waiting for tests to complete");
-}, 1000 * 60 * 60);
